pyproject.toml
[project]
name = "FlatProt"
version = "1.0.0"
description = "Python package for creating simplified 2D protein visualizations. Specialized on showing differences and similarities between proteins of the same protein family."
readme = "README.md"
license = "Apache-2.0"

authors = [
    { name = "Tobias O", email = "tobias.olenyi@tum.de" },
    { name = "Constantin C", email = "ge37goy@mytum.de" },
]
requires-python = ">=3.11,<=3.13"
dependencies = [
    "drawsvg>=2.4.0",
    "gemmi>=0.7.0",
    "polars>=1.17.1",
    "pydantic-extra-types>=2.10.1",
    "pydantic>=2.10.3",
    "pydantic-settings>=2.6.1",
    "tqdm>=4.67.1",
    "cyclopts>=3.9.0",
    "toml>=0.10.2",
    "h5py>=3.13.0",
    "numpydantic>=1.6.8",
]

[project.scripts]
flatprot = "flatprot.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "jupyterlab>=4.3.2",
    "mkdocs>=1.6.1",
    "pytest-mock>=3.14.0",
    "pytest>=8.3.4",
    "jupytext>=1.17.0",
    "mkdocstrings[python]>=0.29.1",
]
db-builder = [
    "httpx-retries>=0.3.0",
    "snakemake>=9.0.1",
]

[project.urls]
Repository = "https://github.com/t03i/FlatProt"
Documentation = "https://t03i.github.io/FlatProt/"
Zenodo = "https://doi.org/10.5281/zenodo.10674045"
Example-Collab = "https://colab.research.google.com/drive/17u0twE81kYYspNFsdXUHrCyP33hj0dO6?usp=sharing"
